+++ Arvore_CPP/Arvore_CPP.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25420.1
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Arvore_CPP", "Arvore_CPP\Arvore_CPP.vcxproj", "{B94549AA-8374-40B5-B3E3-85D4C439C734}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Debug|x64.ActiveCfg = Debug|x64
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Debug|x64.Build.0 = Debug|x64
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Debug|x86.ActiveCfg = Debug|Win32
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Debug|x86.Build.0 = Debug|Win32
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Release|x64.ActiveCfg = Release|x64
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Release|x64.Build.0 = Release|x64
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Release|x86.ActiveCfg = Release|Win32
+		{B94549AA-8374-40B5-B3E3-85D4C439C734}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
