+++ Arvore_CPP/Arvore_CPP/Arvore.cpp
@@ -1,0 +1,5 @@
+// Arvore.cpp : Defines the entry point for the console application.
+//
+
+#include "stdafx.h"
+
