+++ Arvore_CPP/Arvore_CPP/NoArvore.cpp
@@ -1,0 +1,2 @@
+#include "stdafx.h"
+#include "NoArvore.h"
