--- Matriz_cpp.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.30501.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Matriz_cpp", "Matriz_cpp\Matriz_cpp.vcxproj", "{3393044F-DAD9-4E97-9CD1-0D105D554CDB}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Release|Win32 = Release|Win32
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3393044F-DAD9-4E97-9CD1-0D105D554CDB}.Debug|Win32.ActiveCfg = Debug|Win32
-		{3393044F-DAD9-4E97-9CD1-0D105D554CDB}.Debug|Win32.Build.0 = Debug|Win32
-		{3393044F-DAD9-4E97-9CD1-0D105D554CDB}.Release|Win32.ActiveCfg = Release|Win32
-		{3393044F-DAD9-4E97-9CD1-0D105D554CDB}.Release|Win32.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
