--- Matriz_cpp/ListaDupla.cpp
@@ -1,1 +1,0 @@
-#include "stdafx.h"
