--- Matriz_cpp/NoLista.cpp
@@ -1,2 +1,0 @@
-#include "stdafx.h"
-#include "NoLista.h"
